Prepare persistence of airport client.
I use a proxy between planes and client actorRefs (atc and groundControl). Like this, when restarted, the plane can contact the same proxy that will route the message to the new remote actor

Renamed UserInfo into SessionInfo.

Change EventListener and Websocket. In case of a long run, server burned in OutOfMemory